Extract Test.Util.Corruption from Test.Consensus.Byron.Ledger
We will reuse it in the Shelley tests.
