+++ ouroboros-consensus-byron/test/Test/Consensus/Byron/Ledger.hs
@@ -74,6 +74,7 @@ import qualified Test.Cardano.Chain.UTxO.Example as CC
 import qualified Test.Cardano.Chain.UTxO.Gen as CC
 import qualified Test.Cardano.Crypto.Gen as CC
 
+import           Test.Util.Corruption
 import           Test.Util.Orphans.Arbitrary ()
 import           Test.Util.Roundtrip
 
@@ -337,52 +338,6 @@ testCfg = pInfoConfig $ protocolInfoByron
     (CC.Update.ProtocolVersion 1 0 0)
     (CC.Update.SoftwareVersion (CC.Update.ApplicationName "Cardano Test") 2)
     Nothing
-
-newtype Corruption = Corruption Word
-  deriving stock   (Show)
-  deriving newtype (Arbitrary)
-
--- | Increment (overflow if necessary) the byte at position @i@ in the
--- bytestring, where @i = n `mod` length bs@.
---
--- If the bytestring is empty, return it unmodified.
-applyCorruption :: Corruption -> Lazy.ByteString -> Lazy.ByteString
-applyCorruption (Corruption n) bs
-    | Lazy.null bs
-    = bs
-    | otherwise
-    = before <> Lazy.cons (Lazy.head atAfter + 1) (Lazy.tail atAfter)
-  where
-    offset = fromIntegral n `mod` Lazy.length bs
-    (before, atAfter) = Lazy.splitAt offset bs
-
--- | Serialise @a@, apply the given corruption, deserialise it, when that
--- fails, the corruption was detected. When deserialising the corrupted
--- bytestring succeeds, pass the deserialised value to the integrity checking
--- function. If that function returns 'False', the corruption was detected, if
--- it returns 'True', the corruption was not detected and the test fails.
-detectCorruption
-  :: Show a
-  => (a -> Encoding)
-  -> (forall s. Decoder s (Lazy.ByteString -> a))
-  -> (a -> Bool)
-     -- ^ Integrity check that should detect the corruption. Return 'False'
-     -- when corrupt.
-  -> a
-  -> Corruption
-  -> Property
-detectCorruption enc dec isValid a cor = case deserialiseFromBytes dec bs of
-    Right (_, mkA')
-        | not (isValid a')
-        -> label "corruption detected" $ property True
-        | otherwise
-        -> label "corruption not detected" $
-           counterexample ("Corruption not detected: " <> show a') False
-      where
-        a' = mkA' bs
-    Left _ -> label "corruption detected by decoder" $ property True
-  where
-    bs = applyCorruption cor $ toLazyByteString (enc a)
 
 {-------------------------------------------------------------------------------
   Generators
