+++ ouroboros-consensus/ouroboros-consensus-test-infra/ouroboros-consensus-test-infra.cabal
@@ -34,6 +34,7 @@ library
                        Test.Util.Blob
                        Test.Util.BlockchainTime
                        Test.Util.ChunkInfo
+                       Test.Util.Corruption
                        Test.Util.FS.Sim.Error
                        Test.Util.FS.Sim.FsTree
                        Test.Util.FS.Sim.MockFS
